+++ opentelemetry-sdk/CHANGELOG.md
@@ -8,6 +8,8 @@
   ([#761](https://github.com/open-telemetry/opentelemetry-python/pull/761))
 - bugfix: byte type attributes are decoded before adding to attributes dict
   ([#775](https://github.com/open-telemetry/opentelemetry-python/pull/775))
+- Rename Observer to ValueObserver
+  ([#764](https://github.com/open-telemetry/opentelemetry-python/pull/764))
 
 ## 0.8b0